+++ packages/deployer/src/utils/mnemonic.ts
